requirements.txt

django-common-helpers==0.9.2
django-cron==0.5.1
django-crontab==0.7.1
django-filter==2.4.0
django-multiselectfield==0.1.12
django-poormanscron==0.1.3
djangorestframework==3.12.4
json5==0.9.5
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
